requirements.txt
requests
art
python-dotenv